requirements.txt
pyspark~=3.5.5
pytest~=8.3.5
chispa~=0.11.1
apache-airflow
apache-airflow-providers-apache-spark==4.1.5
